+++ linuxsampler/trunk/src/engines/gig/Voice.cpp
@@ -4,7 +4,8 @@
  *                                                                         *
  *   Copyright (C) 2003, 2004 by Benno Senoner and Christian Schoenebeck   *
  *   Copyright (C) 2005 - 2008 Christian Schoenebeck                       *
- *   Copyright (C) 2009 - 2015 Christian Schoenebeck and Grigor Iliev      *
+ *   Copyright (C) 2009 Christian Schoenebeck and Grigor Iliev             *
+ *   Copyright (C) 2010 - 2016 Christian Schoenebeck and Andreas Persson   *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -58,7 +59,7 @@ namespace LinuxSampler { namespace gig {
         si.ChannelCount     = pSample->Channels;
         si.FrameSize        = pSample->FrameSize;
         si.BitDepth         = pSample->BitDepth;
-        si.TotalFrameCount  = pSample->SamplesTotal;
+        si.TotalFrameCount  = (uint)pSample->SamplesTotal;
 
         si.HasLoops       = pRegion->SampleLoops;
         si.LoopStart      = (si.HasLoops) ? pRegion->pSampleLoops[0].LoopStart  : 0;
